Fix configuration for database.yml file
Previously, the Dockerfile that existed was not appropriate for
deployment because there was no way to customize the database
configuration once the container was built.

Now, by using ENV variables, we can inject our database configuration
into this file at our leisure.

This commit also fixes an error that existed inside
`build/dockerfile_writer` regarding using a `~` with a HEREDOC.